bitmoji_credit_app/requirements-fastapi.txt
# 5 Minutes to Credit Wellness — FastAPI Dependencies
# AE Labs — (c) 2025 Sean Gilmore / Arden Edge Capital

# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36

# Auth
argon2-cffi==23.1.0
PyJWT==2.9.0

# Encryption
cryptography==42.0.5

# PDF parsing
pdfplumber==0.11.0
pytesseract==0.3.13
Pillow>=12.0.0
python-docx==1.1.2
beautifulsoup4==4.12.3

# PDF generation
reportlab==4.2.5

# Payments
stripe==8.4.0

# HTTP
requests==2.31.0

# Utils
python-dotenv==1.0.1
werkzeug==3.0.1
